requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pymongo==4.5.0
python-multipart==0.0.6
pillow==10.0.1
numpy==1.24.3
pydantic==2.3.0
opencv-python==4.7.0.72
python-jose==3.3.0
uuid==1.30
aiofiles==23.1.0
apscheduler==3.10.4
insightface==0.7.3
onnxruntime==1.19.1
gdown==4.7.1